+++ ChoreoHelper.ViewModels/DistanceViewModel.cs
@@ -7,6 +7,6 @@ namespace ChoreoHelper.ViewModels;
 [DebuggerDisplay("{Description}")]
 public sealed class DistanceViewModel : ReactiveObject
 {
-    [Reactive] public OneOf<byte, None> Distance { get; set; } = new None();
+    [Reactive] public OneOf<float, None> Distance { get; set; } = new None();
     [Reactive] public string Description { get; set; } = string.Empty;
 }
